add roles and users

+++ routes/web.php
@@ -13,39 +13,41 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-use Spatie\Permission\Models\Role;
-use Spatie\Permission\Models\Permission;
+// use Spatie\Permission\Models\Role;
+// use Spatie\Permission\Models\Permission;
 
-Route::get('/roles', function () {
-    $roleAdmin = Role::create(['name' => 'Administrador']);
-    $roleCordi = Role::create(['name' => 'Coordinador']);
+// Route::get('/roles', function () {
+//     // $roleAdmin = Role::create(['name' => 'Administrador']);
+//     // $roleCordi = Role::create(['name' => 'Coordinador']);
+//     $roleDev = Role::create(['name' => 'Developer']);
 
-    // $permission = Permission::create(['name' => 'edit articles']);
-    $userAdmin = \App\Models\User::find(1);
-    $userAdmin->assignRole('Administrador');
-    echo ('Role Admin Ok');
+//     // $permission = Permission::create(['name' => 'edit articles']);
+//     $userDev = \App\Models\User::find(1);
+//     $userDev->assignRole('Developer');
+//     echo ('Role Developer Ok');
 
-    $userCordi = \App\Models\User::find(2);
-    $userCordi->assignRole('Coordinador');
-    echo ('Role Cordi Ok');
-});
+//     $userAdmin = \App\Models\User::find(3);
+//     $userAdmin->assignRole('Administrador');
+//     echo ('Role Admin Ok');
+
+//     $userCordi = \App\Models\User::find(4);
+//     $userCordi->assignRole('Coordinador');
+//     echo ('Role Cordi Ok');
+// });
 
 Route::get('/', function () {
     return view('welcome');
 })->name('welcome');
 
-// Route::middleware(['auth:sanctum', 'verified'])->get('/dashboard', function () {
-//     return view('dashboard');
-// })->name('dashboard');
 Route::middleware(['auth'])->group(function () {
 
     Route::view('/dashboard', 'dashboard')->name('dashboard');
 
     Route::group(['prefix' => 'registros', 'controller' => App\Http\Controllers\RegistroDocumentoController::class], function () {
         Route::get('/', 'index')->name('registro.index');
-        Route::get('/create', 'create')->name('registro.create');
-        Route::post('/store', 'store')->name('registro.store');
-        Route::get('/show/{registro}', 'show')->name('registro.show');
-        Route::delete('/destroy/{registro}', 'destroy')->name('registro.destroy');
+        Route::get('/create', 'create')->name('registro.create')->middleware(['role:Coordinador|Developer']);
+        Route::post('/store', 'store')->name('registro.store')->middleware(['role:Coordinador|Developer']);
+        Route::get('/show/{registro}', 'show')->name('registro.show')->middleware(['role:Administrador|Developer']);
+        Route::delete('/destroy/{registro}', 'destroy')->name('registro.destroy')->middleware(['role:Administrador|Developer']);
     });
 });
